% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readers.R
\name{clamp_h5_schema}
\alias{clamp_h5_schema}
\title{Expected HDF5 schema for CLAMP expression files}
\usage{
clamp_h5_schema()
}
\value{
A data frame with columns \code{path}, \code{type}, and
  \code{description}.
}
\description{
Returns a data frame of the paths, types, and purpose of each dataset
required in a CLAMP HDF5 expression file.
}
\examples{
clamp_h5_schema()

}
