% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\name{human_gene_v2_5_alz_h5}
\alias{human_gene_v2_5_alz_h5}
\title{human_gene_v2_5_alz_h5}
\usage{
human_gene_v2_5_alz_h5()
}
\value{
A character string representing the local file path to the downloaded resource.
}
\description{
Retrieves the H5 file (path) containing gene sets and pathway priors from ExperimentHub.
}
\examples{
\dontrun{
  h5_path <- CLAMPData::human_gene_v2_5_alz_h5()
  # Use rhdf5::h5read to access the file
}
}
