% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\docType{data}
\name{GSE164416_DP_htseq_counts_txt_gz}
\alias{GSE164416_DP_htseq_counts_txt_gz}
\title{GSE164416_DP_htseq_counts_txt_gz}
\format{
## `GSE164416_DP_htseq_counts_txt_gz()`
A data frame with x,xxx rows and xx columns:
\describe{
  \item{field1}{describe}
  \item{field2}{describe}
  \item{field3}{describe}
  ...
}
}
\source{
<https://zenodo.org/records/17014313/files/GSE164416_DP_htseq_counts.txt.gz?download=1>
}
\usage{
GSE164416_DP_htseq_counts_txt_gz
}
\description{
A brief description of this dataset
}
\examples{
data(GSE164416_DP_htseq_counts_txt_gz)
head(GSE164416_DP_htseq_counts_txt_gz)
}
\keyword{datasets}
