% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\name{GSE164416_DP_htseq_counts_txt_gz}
\alias{GSE164416_DP_htseq_counts_txt_gz}
\title{GSE164416_DP_htseq_counts_txt_gz}
\usage{
GSE164416_DP_htseq_counts_txt_gz()
}
\value{
A character string representing the local file path to the downloaded resource.
}
\description{
Retrieves the raw count matrix (path to gzipped file) for the GSE164416 dataset from ExperimentHub.
}
\examples{
counts_path <- CLAMPData::GSE164416_DP_htseq_counts_txt_gz()
}
