% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\docType{data}
\name{human_gene_v2_5_alz_h5}
\alias{human_gene_v2_5_alz_h5}
\title{human_gene_v2_5_alz_h5}
\format{
## `human_gene_v2_5_alz_h5()`
A data frame with x,xxx rows and xx columns:
\describe{
  \item{field1}{describe}
  \item{field2}{describe}
  \item{field3}{describe}
  ...
}
}
\source{
<https://zenodo.org/records/17014313/files/human_gene_v2.5_alz.h5?download=1>
}
\usage{
human_gene_v2_5_alz_h5
}
\description{
A brief description of this dataset
}
\examples{
data(human_gene_v2_5_alz_h5)
head(human_gene_v2_5_alz_h5)
}
\keyword{datasets}
