% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\name{islets_metadata_csv}
\alias{islets_metadata_csv}
\title{islets_metadata_csv}
\usage{
islets_metadata_csv()
}
\value{
A character string representing the local file path to the downloaded resource.
}
\description{
Retrieves the metadata CSV file (path) for the islets example from ExperimentHub.
}
\examples{
\dontrun{
  metadata_path <- CLAMPData::islets_metadata_csv()
  # You can then use read.csv to load the data
}
}
