% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\docType{data}
\name{islets_metadata_csv}
\alias{islets_metadata_csv}
\title{islets_metadata_csv}
\format{
## `islets_metadata_csv()`
A data frame with x,xxx rows and xx columns:
\describe{
  \item{field1}{describe}
  \item{field2}{describe}
  \item{field3}{describe}
  ...
}
}
\source{
<https://zenodo.org/records/17014313/files/islets_metadata.csv?download=1>
}
\usage{
islets_metadata_csv
}
\description{
A brief description of this dataset
}
\examples{
data(islets_metadata_csv)
head(islets_metadata_csv)
}
\keyword{datasets}
